Minor updates to regard-updated-approval-status
