+++ server/application-data/application-data.json
@@ -1,4 +1,1 @@
-{
-    "application-registration-time": [],
-    "preceeding-application-information": []
-}
+{"application-registration-time":[],"preceeding-application-information":[]}
