+++ server/database/config.json
@@ -5110,7 +5110,7 @@
                         "name": [
                             {
                                 "value-name": "ForwardingKind",
-                                "value": "core-model-1-4:FORWARDING_KIND_TYPE_INVARIANT_PROCESS_SNIPPET"
+                                "value": "core-model-1-4:FORWARDING_KIND_TYPE_PROCESS_SNIPPET"
                             },
                             {
                                 "value-name": "ForwardingName",
@@ -5205,7 +5205,7 @@
                         "name": [
                             {
                                 "value-name": "ForwardingKind",
-                                "value": "core-model-1-4:FORWARDING_KIND_TYPE_INVARIANT_PROCESS_SNIPPET"
+                                "value": "core-model-1-4:FORWARDING_KIND_TYPE_PROCESS_SNIPPET"
                             },
                             {
                                 "value-name": "ForwardingName",
